Set HOME instead of using 'sudo -i' in CMD
'sudo -i' runs the shell specified by the target user's password
database entry as a login shell. This executes the .profile in /root but
this includes a call to mesg, which assumes a TTY is present. This makes
a "default: stdin: is not a tty" appear in the output of the
initialization of the container on Eclipse Che.